Add base WebSocket code
Skeleton code has been provided for movement and CO WebSocket messages.
Movement messages simply cause the server to print the direction.
CO messages are generated whenever the client asks for them. The server
determines if it needs to send a toggle back to the indicator light. For
now it always toggles.